api/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
easyocr
opencv-python-headless
google-genai
pydantic
python-dotenv
pytesseract
pillow
pymupdf
composio-core
numpy
